requirements.txt
certifi>=2018.1.18
chardet>=3.0.4
dnspython>=1.15.0
idna>=2.6
pprint>=0.1
requests>=2.18.4
requests-toolbelt>=0.8.0
six>=1.11.0
tabulate>=0.8.2
tetpyclient>=1.0.6
urllib3>=1.22
